requirements/development.txt
py==1.4.31
pymongo==3.3.0
pytest==2.9.2
PyYAML==3.11
tornado==4.4
